--- src/main/java/com/example/jobgsm/domain/email/exception/MemberNotFoundException.java
@@ -1,14 +1,0 @@
-package com.example.jobgsm.domain.email.exception;
-
-import com.example.jobgsm.global.exception.ErrorCode;
-import lombok.Getter;
-
-@Getter
-public class MemberNotFoundException extends RuntimeException{
-    private final ErrorCode errorCode;
-
-    public MemberNotFoundException(String message) {
-        super(message);
-        this.errorCode = ErrorCode.EXPIRE_EMAIL_CODE;
-    }
-}
