--- src/main/java/com/example/jobgsm/domain/email/presentation/dto/request/EmailSentDto.java
@@ -1,22 +1,0 @@
-package com.example.jobgsm.domain.email.presentation.dto.request;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@NoArgsConstructor
-public class EmailSentDto {
-    @Email
-    @NotBlank(message = "이메일(필수)")
-    private String email;
-
-    @JsonCreator
-    public EmailSentDto(@JsonProperty("email") String email) {
-        this.email = email;
-    }
-}
