--- src/main/java/com/example/jobgsm/domain/email/repository/EmailAuthRepository.java
@@ -1,8 +1,0 @@
-package com.example.jobgsm.domain.email.repository;
-
-import com.example.jobgsm.domain.email.entity.EmailAuth;
-import org.springframework.data.repository.CrudRepository;
-
-public interface EmailAuthRepository extends CrudRepository<EmailAuth, String> {
-}
-
