+++ src/main/java/com/example/jobgsm/domain4/email/exception/AuthCodeExpiredException.java
@@ -1,0 +1,14 @@
+package com.example.jobgsm.domain4.email.exception;
+
+import com.example.jobgsm.global4.exception.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class AuthCodeExpiredException extends RuntimeException{
+    private final ErrorCode errorCode;
+
+    public AuthCodeExpiredException(String message) {
+        super(message);
+        this.errorCode = ErrorCode.EXPIRE_EMAIL_CODE;
+    }
+}
