+++ src/main/java/com/example/jobgsm/domain4/email/exception/ManyRequestEmailAuthException.java
@@ -1,0 +1,15 @@
+package com.example.jobgsm.domain4.email.exception;
+
+import com.example.jobgsm.global4.exception.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class ManyRequestEmailAuthException extends RuntimeException{
+
+    private final ErrorCode errorCode;
+
+    public ManyRequestEmailAuthException(String message) {
+        super(message);
+        this.errorCode = ErrorCode.MANY_REQUEST_EMAIL_AUTH;
+    }
+}
