+++ src/main/java/com/example/jobgsm/domain4/email/exception/MisMatchAuthCodeException.java
@@ -1,0 +1,15 @@
+package com.example.jobgsm.domain4.email.exception;
+
+import com.example.jobgsm.global4.exception.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class MisMatchAuthCodeException extends RuntimeException{
+
+    private final ErrorCode errorCode;
+
+    public MisMatchAuthCodeException(String message) {
+        super(message);
+        this.errorCode = ErrorCode.MISMATCH_AUTH_CODE;
+    }
+}
