+++ src/main/java/com/example/jobgsm/domain4/email/repository/EmailAuthRepository.java
@@ -1,0 +1,8 @@
+package com.example.jobgsm.domain4.email.repository;
+
+import com.example.jobgsm.domain4.email.entity.EmailAuth;
+import org.springframework.data.repository.CrudRepository;
+
+public interface EmailAuthRepository extends CrudRepository<EmailAuth, String> {
+}
+
