--- src/main/java/com/example/jobgsm/global/config/SecurityConfig.java
@@ -1,40 +1,0 @@
-package com.example.jobgsm.global.config;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-
-
-@EnableWebSecurity
-@RequiredArgsConstructor
-@Configuration
-public class SecurityConfig {
-
-    @Bean
-    protected SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                .authorizeRequests()
-                .antMatchers( "/login", "/singUp","/email").permitAll()
-                .and()
-                .formLogin()
-                .loginPage("/login")
-                .loginProcessingUrl("/login_proc")
-                .defaultSuccessUrl("/login")
-                .failureUrl("/login")
-                .and()
-                .csrf().disable();
-
-        return http.build();
-
-    }
-    @Bean
-    public PasswordEncoder passwordEncoder(){
-        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
-    }
-}
