--- src/main/java/com/example/jobgsm/global/exception/ErrorCode.java
@@ -1,18 +1,0 @@
-package com.example.jobgsm.global.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum ErrorCode {
-
-    MANY_REQUEST_EMAIL_AUTH("15분에 최대 3번의 이메일 요청만 가능합니다." , 429),
-    EXPIRE_EMAIL_CODE("이메일 인증번호 시간이 만료되었습니다.", 401),
-    MEMBER_NOT_FOUND("존재하지 않는 회원입니다.", 404),
-    MISMATCH_AUTH_CODE("인증번호가 일치하지 않습니다." , 400),
-    ALREADY_EXIST_EMAIL("이미 존재하는 이메일입니다." , 400);
-
-    private final String message;
-    private final int status;
-}
