--- src/main/java/com/example/jobgsm/global/exception/ErrorResponse.java
@@ -1,11 +1,0 @@
-package com.example.jobgsm.global.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ErrorResponse {
-    private final String message;
-    private final int status;
-}
