+++ src/main/java/com/example/jobgsm/global4/exception/ErrorResponse.java
@@ -1,0 +1,11 @@
+package com.example.jobgsm.global4.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ErrorResponse {
+    private final String message;
+    private final int status;
+}
